create card class
using nested enum
